Add new "--plaintext-metadata" flag.
When this flag is used, we don't templatize metadata JSON object for
every single request and json load it, but we only do that once and
re-use cached version for subsequent requests.

This results in much less overhead when metadata template functionality
is not needed and utilizing large metadata objects.