Fix racy requester shutdown logic
This change resolves the panic in #268. The requester shutdown logic was not
correctly waiting for the worker ticker to shutdown which could cause
the shutdown logic to run concurrently with new workers being created.
This doesn't fail race tests because the workers array is protected by
a mutex, but the length of `b.workers` used for `b.workers[i].Stop()`
is not guaranteed to be the same as the length used to collect errors
from errC.

This is possible for example with the default concurrency schedule, and
a service that completes `-c` requests faster than it takes to create
`-c` workers. This means that the load test finishes before the first
worker `TickValue` has been processed. The workers that are created
after the `b.workers[i].Stop()` loop run forever and the process doesn't
respond to SIGINT, it has to be SIGKILLed.